+++ lib/build.gradle
@@ -37,6 +37,7 @@ dependencies {
     compile 'com.google.code.gson:gson:2.6.2'
     compile 'com.auth0:auth0:0.4.0'
     compile 'com.squareup:otto:1.3.8'
+    compile project(":oauth2")
     testCompile 'junit:junit:4.12'
     testCompile 'org.hamcrest:hamcrest-library:1.3'
     testCompile 'org.robolectric:robolectric:3.0'
