+++ lib/src/main/java/com/auth0/android/lock/LockActivity.java
@@ -57,11 +57,6 @@ import com.auth0.android.lock.events.DatabaseSignUpEvent;
 import com.auth0.android.lock.events.EnterpriseLoginEvent;
 import com.auth0.android.lock.events.FetchApplicationEvent;
 import com.auth0.android.lock.events.SocialConnectionEvent;
-import com.auth0.android.lock.provider.AuthCallback;
-import com.auth0.android.lock.provider.AuthProvider;
-import com.auth0.android.lock.provider.AuthorizeResult;
-import com.auth0.android.lock.provider.CallbackHelper;
-import com.auth0.android.lock.provider.OAuth2WebAuthProvider;
 import com.auth0.android.lock.provider.ProviderResolverManager;
 import com.auth0.android.lock.utils.ActivityUIHelper;
 import com.auth0.android.lock.utils.Strategies;
@@ -72,6 +67,11 @@ import com.auth0.authentication.AuthenticationAPIClient;
 import com.auth0.authentication.result.Credentials;
 import com.auth0.authentication.result.DatabaseUser;
 import com.auth0.callback.BaseCallback;
+import com.auth0.oauth2.AuthCallback;
+import com.auth0.oauth2.AuthProvider;
+import com.auth0.oauth2.AuthorizeResult;
+import com.auth0.oauth2.CallbackHelper;
+import com.auth0.oauth2.OAuth2WebAuthProvider;
 import com.squareup.okhttp.OkHttpClient;
 import com.squareup.otto.Bus;
 import com.squareup.otto.Subscribe;
