+++ lib/src/main/java/com/auth0/android/lock/provider/AuthProviderResolver.java
@@ -28,6 +28,9 @@ import android.content.Context;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
+import com.auth0.oauth2.AuthCallback;
+import com.auth0.oauth2.AuthProvider;
+
 public interface AuthProviderResolver {
 
     /**
