+++ lib/src/main/res/values/strings.xml
@@ -25,11 +25,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools" tools:keep="@string/com_auth0_lock_social_*">
     <!-- Social Authentication -->
     <string name="com_auth0_lock_social_error_authentication">Error parsing Authentication data</string>
-    <string name="com_auth0_lock_social_error_title">There was an error during authentication</string>
-    <string name="com_auth0_lock_social_access_denied_message">Permissions were not granted. Try again.</string>
-    <string name="com_auth0_lock_social_invalid_authorize_url">Auth0 authorize url not properly set.</string>
-    <string name="com_auth0_lock_social_error_message">The application isn\'t configured properly for the social connection. Please check your Auth0\'s application configuration</string>
-    <string name="com_auth0_lock_social_invalid_state">The received URI state is invalid.</string>
 
     <!-- Social Buttons -->
     <string name="com_auth0_lock_social_unknown_placeholder">LOG IN WITH %s</string>
@@ -101,7 +96,6 @@
     <string name="com_auth0_lock_social_icon_dwolla">D</string>
 
     <!-- Lock General -->
-    <string name="com_auth0_lock_webview_loading_retry">Retry</string>
     <string name="com_auth0_lock_default_country_name_fallback">US</string>
     <string name="com_auth0_lock_default_country_code_fallback">+1</string>
     <string name="com_auth0_lock_header_title">Auth0</string>
@@ -158,11 +152,8 @@
     <string name="com_auth0_lock_passwordless_code_request_error_message">There was an error sending the code</string>
     <string name="com_auth0_lock_passwordless_link_request_error_message">There was an error sending the link</string>
     <string name="com_auth0_lock_passwordless_login_error_invalid_credentials_message">Wrong identity or passcode</string>
-    <string name="com_auth0_lock_permission_missing_title">Missing permissions</string>
-    <string name="com_auth0_lock_permission_missing_description">Some permissions required by this provider were not granted. You can try to authenticate again or go to the application\'s permission screen in the phone settings and grant them. The missing permissions are:\n%s</string>
     <string name="com_auth0_lock_db_login_error_invalid_mfa_code_message">The code is invalid or has expired</string>
     <string name="com_auth0_lock_title_mfa_input_code">Two Step Verification</string>
     <string name="com_auth0_lock_description_mfa_input_code">Please enter a verification code from\n your code generator application.</string>
-
 
 </resources>
