+++ oauth2/build.gradle
@@ -1,5 +1,5 @@
 /*
- * ProviderHandler.java
+ * build.gradle
  *
  * Copyright (c) 2016 Auth0 (http://auth0.com)
  *
@@ -22,25 +22,33 @@
  * THE SOFTWARE.
  */
 
-package com.auth0.android.lock.app;
+apply plugin: 'com.android.library'
 
-import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+android {
+    compileSdkVersion 23
+    buildToolsVersion "23.0.2"
 
-import com.auth0.android.lock.provider.AuthProviderResolver;
-import com.auth0.oauth2.AuthCallback;
-import com.auth0.oauth2.AuthProvider;
-
-public class AuthProviderHandler implements AuthProviderResolver {
-
-    @Nullable
-    @Override
-    public AuthProvider onAuthProviderRequest(Context context, @NonNull AuthCallback callback, @NonNull String connectionName) {
-        AuthProvider provider = null;
-        if (connectionName.equals("google-oauth2")) {
-            provider = new CustomWebAuthProvider(callback);
-        }
-        return provider;
+    defaultConfig {
+        minSdkVersion 15
+        targetSdkVersion 23
+        versionCode 1
+        versionName "1.0"
     }
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+}
+
+dependencies {
+    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile 'com.android.support:appcompat-v7:23.3.0'
+    compile 'com.auth0:auth0:0.4.0'
+    testCompile 'junit:junit:4.12'
+    testCompile 'org.hamcrest:hamcrest-library:1.3'
+    testCompile 'org.robolectric:robolectric:3.0'
+    testCompile 'org.mockito:mockito-core:1.10.19'
+
 }
