+++ oauth2/src/main/res/layout/com_auth0_lock_activity_web_view.xml
@@ -1,7 +1,7 @@
 <!--
-  ~ activity_web_view.xml
+  ~ com_auth0_lock_activity_web_view.xml
   ~
-  ~ Copyright (c) 2014 Auth0 (http://auth0.com)
+  ~ Copyright (c) 2016 Auth0 (http://auth0.com)
   ~
   ~ Permission is hereby granted, free of charge, to any person obtaining a copy
   ~ of this software and associated documentation files (the "Software"), to deal
@@ -26,7 +26,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.auth0.android.lock.provider.WebViewActivity">
+    tools:context="com.auth0.oauth2.WebViewActivity">
 
     <WebView
         android:id="@+id/com_auth0_lock_webview"
@@ -61,8 +61,8 @@
             android:id="@+id/com_auth0_lock_retry"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_centerHorizontal="true"
             android:layout_below="@id/com_auth0_lock_text"
+            android:layout_centerHorizontal="true"
             android:layout_marginTop="10dp"
             android:text="@string/com_auth0_lock_webview_loading_retry" />
