+++ settings.gradle
@@ -1,4 +1,4 @@
-include ':app'
+include ':app', ':oauth2'
 
 include ':lock'
 project(':lock').projectDir = new File(rootProject.projectDir, '/lib')
